+++ characters.py
@@ -3,122 +3,56 @@
 # Date created: 2021-10-24
 # Description: Stores character data to be imported into "main.py"
 # ==============================================================
-# YYYY-MM-DD / CHANGELOG (LATEST AT THE TOP)
-# ==============================================================
-# 2021-11-16 / Modified "characters.py" (this file)
-# 2021-10-26 / Modified "characters.py" (this file)
-# 2021-10-25 / Modified "characters.py" (this file)
-# 2021-10-24 / Created "characters.py" (this file)
-# ==============================================================
-
-# Import statement.
-try:
-    from utility import *
-except:
-    raise Exception("Unable to import files.")
-
-p_class = "rookie"  # Default class.
 
 
-# Dictionary of all character classes.
-dict_class = {"rookie": {
-                        "description": "the classic new recruit",
-                        "bonus": "+5 luck",
-                        "health": 100,
-                        "armor": 10,
-                        "stamina": 100,
-                        "critical rate": 10,
-                        "luck": 5},
-              "close-quarters expert": {
-                        "description":
-                        "slaughter your opponents with brute force",
-                        "bonus": "+10% melee damage",
-                        "health": 120,
-                        "armor": 20,
-                        "stamina": 100,
-                        "critical rate": 15,
-                        "luck": 0},
-              "triggerman": {
-                        "description": "overwhelm your foes with gunfire",
-                        "bonus": "+10% gun damage",
-                        "health": 100,
-                        "armor": 10,
-                        "stamina": 100,
-                        "critical rate": 5,
-                        "luck": 0}
-              }
+class user:
+    """Stores/Manages data related to the user"""
+    def __init__(self, name=None, inventory=[], x=1, y=5, user_pos=0):
+        self.name = name
+        self.inv = inventory
+        self.x = x
+        self.y = y
+        self.user_pos = user_pos
 
+    def enter_uname(self, wipe):
+        while 1:
+            abc = input("What is your name? ")
+            if abc.isalnum() is True:
+                self.name = abc
+                break
+            else:
+                wipe()
+                print("Invalid characters.")
 
-def menu_class():
-    """States the players'current status."""
-    global p_class
-    # Sets title.
-    title = "PLAYER INFO:"
-    # Sets pre_text.
-    pre = f"Current: {p_class}\n"
-    # Prints user's current class specifications (added to pre_text).
-    tag = {"health": " HP",
-           "armor": "AMR",
-           "stamina": "STA",
-           "critical rate": " CR",
-           "luck": " LK"
-           }
-    text = ""
-    for value in dict_class[p_class]:
-        if value == 'description':
-            text = text + f"Description: {dict_class[p_class]['description'].capitalize()}.\nPerk: {dict_class[p_class]['bonus']}.\n"
-        elif value == 'bonus':
-            pass
-        else:
-            text = text + f"{tag[value]}-{dict_class[p_class][value]}\n"
-    text = text[:-1]
-    # Sets options.
-    options = ["change class", "exit"]
-    # Calls ft().
-    input = ft(title=title, pre_text=pre + text, prompt=options)
-    if input == "exit":
-        return "main_menu", None
-    return input, None
+    def get_name(self):
+        return self.name
 
+    def get_inv(self):
+        return self.inv
 
-def menu_class_change():
-    """Menu for changing player's class"""
-    global p_class
-    # Sets title.
-    title = "PICK YOUR CLASS:"
-    # Sets list of all available character classes.
-    options = []
-    for class_item in dict_class:
-        options.append(class_item)
-    options.append("exit")
-    # Calls ft().
-    input = ft(title=title, prompt=options)
-    # Changes user's current class to the selected class if applicable.
-    if input == "exit":
-        return "status", None
-    else:
-        p_class = input
-        return "status", "Class Changed!"
+    def add_inv(self, item):
+        self.inv.append(item)
 
+    def rem_inv(self, item):
+        self.inv.remove(item)
 
- 
+    def get_cor(self):
+        return (self.x, self.y)
 
-def menu_class_change():
-    """Menu for changing player's class"""
-    global p_class
-    # Sets title.
-    title = "PICK YOUR CLASS:"
-    # Sets list of all available character classes.
-    options = []
-    for class_item in dict_class:
-        options.append(class_item)
-    options.append("exit")
-    # Calls ft().
-    input = ft(title=title,prompt=options)
-    # Changes user's current class to the selected class if applicable.
-    if input == "exit":
-        return "status", None
-    else:
-        p_class = input
-        return "status", "Class Changed!"
+    def cha_cor(self, plane, value):
+        if value not in (-1, 1):
+            raise Exception("Invalid Value (cha_cor)")
+        if plane == "x":
+            self.x += value
+        elif plane == "y":
+            self.y += value
 
+    def set_cor(self, x, y):
+        self.x = x
+        self.y = y
+
+    def get_upos(self):
+        return self.user_pos
+
+    def cha_upos(self, value):
+        self.user_pos = value